refactored utils/utils.ts
